+++ src/com/kaylerrenslow/armaDialogCreator/arma/control/ArmaControlRenderer.java
@@ -143,7 +143,9 @@ public class ArmaControlRenderer extends SimpleCanvasComponent implements Viewpo
 	 this whenever a new render needs to happen.
 	 */
 	public final void requestRender() {
-		myControl.getRenderUpdateGroup().update(this.getMyControl());
+		if (myControl.getDisplay() != null) {
+			myControl.getRenderUpdateGroup().update(this.getMyControl());
+		}
 	}
 
 	/**