+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/treeview/dataCreator/ButtonDataCreator.java
@@ -24,7 +24,8 @@ public class ButtonDataCreator implements TreeItemDataCreator<ArmaControl, TreeI
 	@NotNull
 	@Override
 	public TreeItemEntry createNew(@NotNull EditableTreeView<ArmaControl, TreeItemEntry> treeView) {
-		NewControlDialog dialog = new NewControlDialog(ControlType.Button, ArmaDialogCreator.getMainWindow().getCanvasView().getBackgroundControlTreeView() == treeView);
+		NewControlDialog dialog = new NewControlDialog(ControlType.Button, ArmaDialogCreator.getMainWindow()
+				.getCanvasView().isBackgroundTreeView(treeView));
 		dialog.show();
 		if (dialog.wasCancelled()) {
 			return null;