+++ src/com/kaylerrenslow/armaDialogCreator/gui/main/treeview/dataCreator/StaticDataCreator.java
@@ -23,7 +23,7 @@ public class StaticDataCreator implements TreeItemDataCreator<ArmaControl, TreeI
 
 	@Override
 	public TreeItemEntry createNew(@NotNull EditableTreeView<ArmaControl, TreeItemEntry> treeView) {
-		NewControlDialog dialog = new NewControlDialog(ControlType.Static, ArmaDialogCreator.getMainWindow().getCanvasView().getBackgroundControlTreeView() == treeView);
+		NewControlDialog dialog = new NewControlDialog(ControlType.Static, ArmaDialogCreator.getMainWindow().getCanvasView().isBackgroundTreeView(treeView));
 		dialog.show();
 		if (dialog.wasCancelled()) {
 			return null;