+++ src/com/kaylerrenslow/armaDialogCreator/util/UpdateGroupListener.java
@@ -15,4 +15,17 @@ public interface UpdateGroupListener<T> {
 	 */
 	void update(@NotNull UpdateListenerGroup<T> group, @Nullable T data);
 
+	/**
+	 Have the {@link UpdateListenerGroup} automatically check if this listener has expired.
+	 If the listener has been expired, the {@link UpdateListenerGroup} will remove it automatically.
+	 The time that it is removed is entirely up to the {@link UpdateListenerGroup}.<p>
+	 <p>
+	 Default return value is false.
+
+	 @return true if expired, false if not expired
+	 */
+	default boolean hasExpired() {
+		return false;
+	}
+
 }